Refactor Darwin file picker utilities for iOS and macOS
- Replaced `FilePickerUtils.swift` with `DarwinFilePickerUtils` to share logic across Apple platforms.
- Introduced `IOSFilePickerUtils` and `MacOSFilePickerUtils` to handle platform-specific content type inference.
- Updated `MacOSFilePickerHandler` to use the utility for generating default filenames in the save dialog.
- Updated `IOSFilePickerHandler` to utilize the new utility structure.
