+++ darwin/file_picker/Sources/file_picker/IOSFilePickerHandler.swift
@@ -259,7 +259,10 @@ final class IOSFilePickerHandler: NSObject,
     private func saveFile(_ arguments: [String: Any]) {
         isSaveFile = true
         let bytes = arguments["bytes"] as? FlutterStandardTypedData
-        let fileName = FilePickerUtils.buildSaveFileName(from: arguments, data: bytes?.data, lastPickedFileExtension: lastPickedFileExtension)
+        let fileName = IOSFilePickerUtils.buildSaveFileName(
+            from: arguments,
+            data: bytes?.data,
+            lastPickedFileExtension: lastPickedFileExtension)
         let tempFile = URL(fileURLWithPath: NSTemporaryDirectory())
             .appendingPathComponent(fileName)
 
@@ -301,8 +304,8 @@ final class IOSFilePickerHandler: NSObject,
         }
     }
 
-    // buildSaveFileName, inferSaveExtension and inferExtensionFromDataSignature
-    // have been moved to FilePickerUtils.swift to avoid duplication across handlers.
+    // The shared save-file-name flow lives in DarwinFilePickerUtils, while
+    // iOS-specific content type inference lives in IOSFilePickerUtils.
 
     private func updateLastPickedFileExtension(from fileURL: URL) {
         let pathExtension = fileURL.pathExtension.trimmingCharacters(in: .whitespacesAndNewlines)
